pull out feature selection
